@andyzhangx 2025-08-07
> Thanks for the PR! Could you do a quick benchmark to show TTFT with and without using LMCache? Benchmarking script(s) you can use: https://github.com/vllm-project/vllm/tree/main/benchmarks

@chewong I have provided the test result in the PR description, thanks.