@andyzhangx 2025-08-07
> could you share some benchmark result? It will guide user to set the utilization for cpu mem cache. Use both random and prefix aware dataset for benchmark.

@zhuangqh I have provided the test result (it's random test) in the PR description, thanks.