@chewong 2025-08-12
@zhuangqh @andyzhangx  Do you know why TTFT is lower when offloading KV cache to CPU memory? If I understand correctly, this feature benefits multi-turn conversations and/or high-throughput environment the most because it requires a larger context window, which can easily exceed GPU memory capacity. However, the benchmark above is only using 256 input tokens, which is a very short context window.

Could we try running some multi-turn benchmark like https://github.com/vllm-project/vllm/tree/main/benchmarks/multi_turn to further verify the hypothesis?